+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_occ.geometry.OCCNurbsSurface.from_extrusion`.
 * Added implementation for `compas_occ.geometry.OCCCurve.divide_by_count`.
 * Added `compas_occ.geometry.OCCCurve.divide` as alias for `compas_occ.geometry.OCCCurve.divide_by_count`.
-* Added `compas_occ.geometry.OCCCurve.project` and `compas_occ.geometry.OCCCurve.projected`.
+* Added `compas_occ.geometry.OCCCurve.projected`.
 * Added `compas_occ.geometry.OCCCurve.embedded`.
 
 ### Changed
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed unused precision parameter in `compas_occ.geometry.OCCCurve.length`.
 * Fixed bug in `compas_occ.brep.BRep.to_meshes`.
 * Changed `compas_frame_from_occ_position` to `compas_frame_from_occ_ax3`.
+* Changed `compas_occ.brep.BRep.to_tesselation` to use `BRepMesh_IncrementalMesh`.
 
 ### Removed